feat: moved tests to tests folder
